requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cognitiveservices-speech
azure-ai-inference
azure-core
azure-identity
azure-storage-blob
openai
flask
flask-cors
requests